added value arc!
needs some refinement, not sold on the colors. i was being lazy to make it work by just reducing the bounds of where i'm drawing the rotary sliders so there's enough room to draw the value arc, but the sliders resize based on window / height width, so the position of the value arc breaks on resizing....

but it works! i've wanted to do this ever since I saw it in deelay